Add rudimentary support for decoding Bignum values
It turns out that even a simple Unix timestamp is enough to cause Ruby
Marshal to output a Bignum. The encoding is reasonably simple: a sign
byte (ASCII + or -), half the byte length, then the Bignum's bytes.

Note that as the math/big package requires a big-endian byte encoding,
we assume that the input is little-endian and reverse it as needed.

For details, see https://docs.ruby-lang.org/en/master/doc/marshal_rdoc.html#label-Bignum.